@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :agency do
     sequence(:name) { |n| "test-agency-#{n}" }
   end
