@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -16,101 +15,89 @@ ActiveRecord::Schema.define(version: 20180309010908) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
-  create_table "agencies", force: :cascade do |t|
-    t.string   "name",       null: false
+  create_table "agencies", id: :serial, force: :cascade do |t|
+    t.string "name", null: false
     t.datetime "created_at"
     t.datetime "updated_at"
+    t.index ["name"], name: "index_agencies_on_name", unique: true
   end
 
-  add_index "agencies", ["name"], name: "index_agencies_on_name", unique: true, using: :btree
-
-  create_table "ar_internal_metadata", primary_key: "key", force: :cascade do |t|
-    t.string   "value"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-  end
-
-  create_table "delayed_jobs", force: :cascade do |t|
-    t.integer  "priority",   default: 0, null: false
-    t.integer  "attempts",   default: 0, null: false
-    t.text     "handler",                null: false
-    t.text     "last_error"
+  create_table "delayed_jobs", id: :serial, force: :cascade do |t|
+    t.integer "priority", default: 0, null: false
+    t.integer "attempts", default: 0, null: false
+    t.text "handler", null: false
+    t.text "last_error"
     t.datetime "run_at"
     t.datetime "locked_at"
     t.datetime "failed_at"
-    t.string   "locked_by"
-    t.string   "queue"
+    t.string "locked_by"
+    t.string "queue"
     t.datetime "created_at"
     t.datetime "updated_at"
+    t.index ["priority", "run_at"], name: "delayed_jobs_priority"
   end
 
-  add_index "delayed_jobs", ["priority", "run_at"], name: "delayed_jobs_priority", using: :btree
-
-  create_table "groups", force: :cascade do |t|
-    t.datetime "created_at",               null: false
-    t.datetime "updated_at",               null: false
-    t.string   "name",                     null: false
-    t.text     "description", default: ""
-  end
-
-  add_index "groups", ["name"], name: "index_groups_on_name", unique: true, using: :btree
-
-  create_table "service_providers", force: :cascade do |t|
-    t.integer  "user_id",                                               null: false
-    t.string   "issuer",                                                null: false
-    t.string   "friendly_name",                                         null: false
-    t.text     "description"
-    t.text     "metadata_url"
-    t.text     "acs_url"
-    t.text     "assertion_consumer_logout_service_url"
-    t.text     "saml_client_cert"
-    t.integer  "block_encryption",                      default: 1,     null: false
-    t.datetime "created_at"
-    t.datetime "updated_at"
-    t.boolean  "active",                                default: true,  null: false
-    t.boolean  "approved",                              default: false, null: false
-    t.text     "sp_initiated_login_url"
-    t.text     "return_to_sp_url"
-    t.integer  "agency_id"
-    t.json     "attribute_bundle"
-    t.integer  "group_id"
-    t.string   "logo"
-    t.integer  "identity_protocol",                     default: 0
-    t.json     "redirect_uris"
-  end
-
-  add_index "service_providers", ["group_id"], name: "index_service_providers_on_group_id", using: :btree
-  add_index "service_providers", ["issuer"], name: "index_service_providers_on_issuer", unique: true, using: :btree
-
-  create_table "user_groups", force: :cascade do |t|
-    t.integer  "user_id"
-    t.integer  "group_id"
+  create_table "groups", id: :serial, force: :cascade do |t|
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
+    t.string "name", null: false
+    t.text "description", default: ""
+    t.index ["name"], name: "index_groups_on_name", unique: true
   end
 
-  add_index "user_groups", ["group_id"], name: "index_user_groups_on_group_id", using: :btree
-  add_index "user_groups", ["user_id"], name: "index_user_groups_on_user_id", using: :btree
-
-  create_table "users", force: :cascade do |t|
-    t.string   "uuid"
-    t.string   "email",                              null: false
-    t.string   "first_name"
-    t.string   "last_name"
+  create_table "service_providers", id: :serial, force: :cascade do |t|
+    t.integer "user_id", null: false
+    t.string "issuer", null: false
+    t.string "friendly_name", null: false
+    t.text "description"
+    t.text "metadata_url"
+    t.text "acs_url"
+    t.text "assertion_consumer_logout_service_url"
+    t.text "saml_client_cert"
+    t.integer "block_encryption", default: 1, null: false
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.integer  "sign_in_count",      default: 0,     null: false
-    t.datetime "current_sign_in_at"
-    t.datetime "last_sign_in_at"
-    t.inet     "current_sign_in_ip"
-    t.inet     "last_sign_in_ip"
-    t.boolean  "admin",              default: false, null: false
-    t.integer  "group_id"
+    t.boolean "active", default: true, null: false
+    t.boolean "approved", default: false, null: false
+    t.text "sp_initiated_login_url"
+    t.text "return_to_sp_url"
+    t.integer "agency_id"
+    t.json "attribute_bundle"
+    t.integer "group_id"
+    t.string "logo"
+    t.integer "identity_protocol", default: 0
+    t.json "redirect_uris"
+    t.index ["group_id"], name: "index_service_providers_on_group_id"
+    t.index ["issuer"], name: "index_service_providers_on_issuer", unique: true
   end
 
-  add_index "users", ["email"], name: "index_users_on_email", unique: true, using: :btree
-  add_index "users", ["group_id"], name: "index_users_on_group_id", using: :btree
-  add_index "users", ["uuid"], name: "index_users_on_uuid", unique: true, using: :btree
+  create_table "user_groups", id: :serial, force: :cascade do |t|
+    t.integer "user_id"
+    t.integer "group_id"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["group_id"], name: "index_user_groups_on_group_id"
+    t.index ["user_id"], name: "index_user_groups_on_user_id"
+  end
+
+  create_table "users", id: :serial, force: :cascade do |t|
+    t.string "uuid"
+    t.string "email", null: false
+    t.string "first_name"
+    t.string "last_name"
+    t.datetime "created_at"
+    t.datetime "updated_at"
+    t.integer "sign_in_count", default: 0, null: false
+    t.datetime "current_sign_in_at"
+    t.datetime "last_sign_in_at"
+    t.inet "current_sign_in_ip"
+    t.inet "last_sign_in_ip"
+    t.boolean "admin", default: false, null: false
+    t.integer "group_id"
+    t.index ["email"], name: "index_users_on_email", unique: true
+    t.index ["group_id"], name: "index_users_on_group_id"
+    t.index ["uuid"], name: "index_users_on_uuid", unique: true
+  end
 
   add_foreign_key "service_providers", "agencies"
   add_foreign_key "service_providers", "groups"
